Refactoring * Commented out DOS functions * Implemented new API for FilesystemHelper(not implemented in functionality) * partitialy replaced old DOS calls with direct calls for filesystembuffer ! not compilable code in commit

+++ source/opentm2/core/utilities/Base64.h
@@ -40,7 +40,7 @@ static inline bool is_base64(unsigned char c) {
     return (isalnum(c) || (c == '+') || (c == '/'));
 }
 
-std::string base64_encode(unsigned char const* bytes_to_encode, unsigned int in_len) {
+std::string base64_encode(unsigned char const* bytes_to_encode, size_t in_len) {
     std::string ret;
     int i = 0;
     int j = 0;
@@ -82,7 +82,7 @@ std::string base64_encode(unsigned char const* bytes_to_encode, unsigned int in_
     return ret;
 
 }
-unsigned char* base64_decode(std::string const& encoded_string, int& out_size) {
+unsigned char* base64_decode(std::string const& encoded_string, size_t& out_size) {
     int in_len = encoded_string.size();
     int i = 0;
     int j = 0;
